+++ lib/libtab.dart
@@ -1,6 +1,7 @@
 library libtab;
 
 export 'chord.dart';
+export 'chord_notes.dart';
 export 'context.dart';
 export 'display/chart_positioning.dart';
 export 'display/chord_display.dart';